to-do: weigthed average where we can include the last index of the layers, have to modify the minimum method and the median method to ignore the last index if it has a 0 as the value, DYNAMIC LEGEND, alphabetize properties, loading cursor, OT LIKE 6740

+++ index2.php
@@ -386,7 +386,7 @@ function getPolygons(){//this is run button
 	hecho = false;
 	var depth = document.getElementById("depth").value;
 	//console.log(depth);
-	depth = parseInt(depth);
+	depth = parseFloat(depth);
 	//console.log(depth);
 
 	app.payload.depth = depth;
